feat(topics): move TopicEvent model to timeline app
