+++ semanticnews/topics/utils/timeline/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class TimelineConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'semanticnews.topics.utils.timeline'
